Handle types with parentheses in signature regex
The signature regex needed a bit of massaging around the first and
last capture groups, which were fighting a bit. We need the first
capture group to be lazy so it doesn't eat a "subroutine" destined for
the second capture group. However, the last capture group is too
greedy if a type contains parentheses (e.g. "character(len=*)"), and
eats most of our expression! Because we don't expect parentheses in
dummy argument names (I think this is the case in the Fortran
specification), we just disallow the final capture group from matching
them, and everything is right again.